_Fixed_ an issue that an HTC phone fails to recognize puzzle frequently.
I adjusted image processing parameters (to adaptive thresholding) to
cover the HTC phone.  The surface cause of the issue was the difference
of the camera characteristics.  The real cause, however, is that I used
a fixed set of processing parameters.  I just adjusted the fixed
parameters, so I may need to adjust them again to cover new phones, as
long as I stick on the fixed set of parameters...